Create AuthClient interface
