+++ yt/go/yt/interface.go
@@ -37,6 +37,7 @@ import (
 	"io"
 
 	"go.ytsaurus.tech/library/go/core/xerrors"
+
 	"go.ytsaurus.tech/yt/go/guid"
 	"go.ytsaurus.tech/yt/go/schema"
 	"go.ytsaurus.tech/yt/go/ypath"
@@ -927,35 +928,6 @@ type AdminClient interface {
 		options *RemoveMemberOptions,
 	) (err error)
 
-	// http:verb:"set_user_password"
-	// http:params:"user","new_password_sha256","current_password_sha256"
-	SetUserPassword(
-		ctx context.Context,
-		user string,
-		newPassword string,
-		currentPassword string,
-		options *SetUserPasswordOptions,
-	) (err error)
-
-	// http:verb:"issue_token"
-	// http:params:"user","password_sha256"
-	IssueToken(
-		ctx context.Context,
-		user string,
-		password string,
-		options *IssueTokenOptions,
-	) (token string, err error)
-
-	// http:verb:"revoke_token"
-	// http:params:"user","password_sha256","token_sha256"
-	RevokeToken(
-		ctx context.Context,
-		user string,
-		password string,
-		token string,
-		options *RevokeTokenOptions,
-	) error
-
 	// http:verb:"add_maintenance"
 	// http:params:"component","address","type","comment"
 	AddMaintenance(
@@ -1536,6 +1508,38 @@ type GenerateTimestampOptions struct{}
 
 type GetInSyncReplicasOptions struct{}
 
+type AuthClient interface {
+
+	// http:verb:"set_user_password"
+	// http:params:"user","new_password_sha256","current_password_sha256"
+	SetUserPassword(
+		ctx context.Context,
+		user string,
+		newPassword string,
+		currentPassword string,
+		options *SetUserPasswordOptions,
+	) (err error)
+
+	// http:verb:"issue_token"
+	// http:params:"user","password_sha256"
+	IssueToken(
+		ctx context.Context,
+		user string,
+		password string,
+		options *IssueTokenOptions,
+	) (token string, err error)
+
+	// http:verb:"revoke_token"
+	// http:params:"user","password_sha256","token_sha256"
+	RevokeToken(
+		ctx context.Context,
+		user string,
+		password string,
+		token string,
+		options *RevokeTokenOptions,
+	) error
+}
+
 type Client interface {
 	CypressClient
 	FileClient
@@ -1568,6 +1572,7 @@ type Client interface {
 	LowLevelSchedulerClient
 
 	AdminClient
+	AuthClient
 
 	QueryTrackerClient
